pyproject.toml
[tool.poetry]
name = "recall-elevenlabs"
version = "0.1.0"
description = "Meeting bot using Recall.ai with ElevenLabs Conversational AI agents"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
websockets = "^12.0"
httpx = "^0.26.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
pyngrok = "^7.0.0"
replicate = "^0.23.0"
pillow = "^10.2.0"
pyyaml = "^6.0.1"
pydub = "^0.25.1"
python-multipart = "^0.0.21"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
black = "^24.1.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
ignore = ["E501"]

